+++ CHANGELOG.md
@@ -47,6 +47,7 @@ Senior Web Developer
 - **(fix)** CSqlDataprovider support #356 (magefad)
 - **(fix)** TbTypeahead - Add attribute autocomplete="off" by default #285
 - **(enh)** Update Redactor to 8.2.3 #386 (magefad)
+- **(fix)** baseID in checkBoxList and radioButtonList can now be customized via htmlOptions, added container support - yii 1.1.13 (fad)
 
 
 ### YiiBooster version 1.0.5